requirements.txt
pytesseract==0.3.9
pillow==9.0.1
opencv-python==4.5.5.64